#116 code reformation and changes in banner update

+++ app/Http/Controllers/ProductColorController.php
@@ -19,43 +19,45 @@ class ProductColorController extends Controller
         // dd($request->all());
         $color = ProductColor::create($request->all());
         return response()->json([
-            'Message' => "Data inserted successfully",
             'data' => $color,
+            'message' => "Data inserted successfully",
             'status' => 200
-        ]);
+        ],200);
     }
-
-
-    public function show($id)
-    {
-        $color = ProductColor::findOrFail($id);
-        return response()->json([
-            'data' => $color,
-            'status' => 200
-        ]);
-    }
-
 
     public function update(ProductColorRequest $request, $id)
     {
-        $color = ProductColor::findOrFail($id);
-        $color->update($request->all());
-        return response()->json([
-            'Message' => "Data updated successfully",
-            'data' => $color,
-            'status' => 200
-        ]);
+        $color = ProductColor::find($id);
+        if ($color) {
+            $color->update($request->all());
+            return response()->json([
+                'data' => $color,
+                'message' => "Data updated successfully",
+                'status' => 200
+            ],200);
+        } else {
+            return response()->json([
+                'message' => "Data not found",
+                'status' => 404
+            ],404);
+        }
     }
-
 
     public function destroy($id)
     {
-        $color = ProductColor::findOrFail($id);
-        $color->delete();
-        return response()->json([
-            'Message' => "Data deleted successfully",
-            'data' => $color,
-            'status' => 200
-        ]);
+        $color = ProductColor::find($id);
+        if ($color) {
+            $color->delete();
+            return response()->json([
+                'data' => $color,
+                'message' => "Data deleted successfully",
+                'status' => 200
+            ],200);
+        }else {
+            return response()->json([
+                'message' => "Data not found",
+                'status' => 404
+            ],404);
+        }
     }
 }
